// src/seaportEx/types.ts
export interface AssetsQueryParams {
    assets?: {
        asset_contract_addresses: string
        token_ids?: string
    }[],
    owner?: string
    limit?: number
    include_orders?: boolean
}

export interface FeesInfo {
    royaltyFeeAddress: string
    royaltyFeePoint: number
    protocolFeePoint?: number
    protocolFeeAddress?: string
}

export interface AssetCollection extends FeesInfo {
    name: string
    symbol: string
    address?: string
    token_id?: string
    schema_name?: string
    nft_version?: string
    created_date?: string
    sell_orders?: any
}